Create DisplayList component to display the list of courses
