+++ client/src/components/Planner/Checklist.js
@@ -2,6 +2,7 @@ import { ListItem, ListItemText, ListSubheader, Typography } from "@mui/material
 import Paper from "@mui/material/Paper";
 import List from "@mui/material/List"
 import { getRequiredCourses } from "../../helper/rotationHelper";
+import DisplayList from "./DisplayList";
 import _union from "lodash/union";
 
 export default function Checklist({ tabInfo, plannedCourses, csvData }) {
@@ -36,35 +37,19 @@ export default function Checklist({ tabInfo, plannedCourses, csvData }) {
 			</Typography>
       <hr/>
       {tabInfo.restricted.length > 0 && 
-        <List sx={{py: 0}}>
-          <ListSubheader> Restricted Courses: </ListSubheader>
-            {tabInfo.restricted.sort().map((a) => {
-              let primary = `Comp Sci ${a}`;
-              let isTakenOrPlanned = 
-                tabInfo.completed.includes(a) || plannedCourses.includes(a);
-              let color = (isTakenOrPlanned) ? "green" : "red";
-              return(
-                <ListItem sx={{py: 0}}>
-                  <ListItemText sx={{py: 0, px: 2, color: color}} primary={primary}/> 
-                </ListItem>
-              );
-            })}
-        </List>
+        <DisplayList 
+          tabInfo={tabInfo}
+          plannedCourses={plannedCourses}
+          subheader={"Restricted Courses"}
+          displayArray={tabInfo.restricted}
+        />
       }
-      <List sx={{py: 0}}>
-        <ListSubheader> Core: </ListSubheader>
-          {coreCourses.sort().map((a) => {
-            let primary = `Comp Sci ${a}`;
-            let isTakenOrPlanned = 
-              tabInfo.completed.includes(a) || plannedCourses.includes(a);
-            let color = (isTakenOrPlanned) ? "green" : "red";
-            return(
-              <ListItem sx={{py: 0}}>
-                <ListItemText sx={{py: 0, px: 2, color: color}} primary={primary}/> 
-              </ListItem>
-            );
-          })}
-      </List>
+      <DisplayList 
+        tabInfo={tabInfo}
+        plannedCourses={plannedCourses}
+        subheader={"Core"}
+        displayArray={coreCourses}
+      />
     </Paper>
   );
 }
